requirements.txt
streamlit
openai-whisper
opencv-python
requests
python-dotenv
Pillow
numpy<2
fpdf2
yt-dlp
